fix(pix3d): int32-wrap the two DivTable products — 3-03 (G3) + B-01 backports
Two 32-bit int products in the pinned Client-Java (01f16088) wrap mod 2^32 but
were left 64-bit in the rev-244 Go port:

- TextureRaster shade-interpolation `var15 = ((arg8-arg7)*DivTable[var16])>>6`
  (Pix3D.java:2076) — already diagnosed as a bug ("pix3d-3-03") and wrapped by
  goscape-client 0700e334 ("audit G3"), which reached 245.2/254/274 but is not
  an ancestor of rev-244. Backport of 0700e33 audit G3.
- GouraudRaster colour-interpolation `var8 = ((arg7-arg6)*DivTable[arg3])>>15`
  (Pix3D.java:889) — wrapped on rev-254 (audit pix3d-B-01); parity backport,
  same masked-but-real precedent as the F2 C-01 port.

Both are masked for all current goldens/icons (overflow needs the delta operand
>= 2^31/32768 = 65_536; the goldens' shade span tops out at 200<<8 = 51_200 and
colour deltas are far smaller). Proof of no output change: all tri goldens +
palette stay GREEN, and the full 2894-icon icondump run is byte-identical
before and after (hash 05198126…).

+++ pkg/jagex2/graphics/pix3d/pix3d.go
@@ -910,7 +910,12 @@ func GouraudRaster(arg0 []int, arg1, arg4, arg5, arg6, arg7 int) {
 			arg1 += arg4
 			arg3 = (arg5 - arg4) >> 2
 			if arg3 > 0 {
-				var8 = ((arg7 - arg6) * DivTable[arg3]) >> 15
+				// Java: Pix3D.java:889 (pin 01f16088) — 32-bit product wraps before
+				// the >>15 (overflows on steep colour deltas over short spans; audit
+				// pix3d-B-01, parity backport of rev-254's wrap). Masked for the tri
+				// goldens/icons: overflow needs |arg7-arg6| >= 2^31/32768 = 65_536,
+				// far above any golden colour delta.
+				var8 = int(int32((arg7-arg6)*DivTable[arg3])) >> 15
 			} else {
 				var8 = 0
 			}
@@ -2090,7 +2095,13 @@ func TextureRaster(arg0 []int, arg1 []int, arg2, arg3, arg4, arg5, arg6, arg7, a
 	} else {
 		if arg6-arg5 > 7 {
 			var16 = (arg6 - arg5) >> 3
-			var15 = ((arg8 - arg7) * DivTable[var16]) >> 6
+			// Java: Pix3D.java:2076 (pin 01f16088) — 32-bit product wraps before
+			// the >>6 (overflows on bright wide textured triangles; audit
+			// pix3d-3-03, backport of 0700e33 audit G3). Masked for the tri
+			// goldens/icons: overflow needs |arg8-arg7|*DivTable[var16] >= 2^31,
+			// i.e. |arg8-arg7| >= 65_536 at DivTable max 32768; the goldens'
+			// shade span (<= 200<<8 = 51_200) stays below it.
+			var15 = int(int32((arg8-arg7)*DivTable[var16])) >> 6
 		} else {
 			var16 = 0
 			var15 = 0
